backend/requirements.txt
fastapi
sqlalchemy
psycopg2-binary
pyjwt
passlib[bcrypt]
gunicorn
python-jose
python-multipart
uvicorn
python-dotenv